More reasonable data types

+++ internal/http_app/dao/chatroom.go
@@ -20,11 +20,11 @@ func (d *chatroom) GetGroupByID(id string, selects ...string) (group model.Group
 }
 
 // * 获取聊天室成员
-func (d *chatroom) GetMembers(chatroom_id string) (member_ids []uint, err error) {
+func (d *chatroom) GetMembers(chatroom_id string) (member_ids []int64, err error) {
 	members, err := Dao.Ris().Do("SMEMBERS", "set_im_chatroom_member:"+chatroom_id)
 	for _, v := range members.([]interface{}) {
 		user_id, _ := strconv.Atoi(string(v.([]uint8)))
-		member_ids = append(member_ids, uint(user_id))
+		member_ids = append(member_ids, int64(user_id))
 	}
 	return
 }
